requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
addict=2.4.0=pypi_0
annotated-doc=0.0.4=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.12.1=pypi_0
argon2-cffi=25.1.0=pypi_0
argon2-cffi-bindings=25.1.0=pypi_0
arrow=1.4.0=pypi_0
asttokens=3.0.1=pypi_0
async-lru=2.0.5=pypi_0
attrs=25.4.0=pypi_0
babel=2.17.0=pypi_0
bcrypt=5.0.0=pypi_0
beautifulsoup4=4.14.3=pypi_0
bleach=6.3.0=pypi_0
bzip2=1.0.8=hda65f42_8
c-ares=1.34.6=hb03c661_0
ca-certificates=2026.1.4=hbd8a1cb_0
cachetools=6.2.4=pypi_0
certifi=2026.1.4=pypi_0
charset-normalizer=3.4.4=pypi_0
click=8.3.1=pypi_0
cloudpickle=3.1.2=pypi_0
comm=0.2.3=pypi_0
contourpy=1.3.3=pypi_0
cryptography=46.0.3=pypi_0
cycler=0.12.1=pypi_0
cython=3.2.4=pypi_0
debugpy=1.8.19=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.4.0=pypi_0
durationpy=0.10=pypi_0
executing=2.2.1=pypi_0
fastapi=0.128.0=pypi_0
fastjsonschema=2.21.2=pypi_0
fiftyone=1.11.0=pypi_0
filelock=3.20.2=pypi_0
fonttools=4.61.1=pypi_0
fsspec=2025.12.0=pypi_0
google-auth=2.47.0=pypi_0
hf-xet=1.2.0=pypi_0
huggingface-hub=0.36.0=pypi_0
icu=75.1=he02047a_0
idna=3.11=pypi_0
imageio=2.37.2=pypi_0
ipykernel=7.1.0=pypi_0
ipython=9.9.0=pypi_0
ipython-pygments-lexers=1.1.1=pypi_0
ipywidgets=8.1.8=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.5.3=pypi_0
json5=0.13.0=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.26.0=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-client=8.8.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.9.1=pypi_0
jupyter-events=0.12.0=pypi_0
jupyter-lsp=2.3.0=pypi_0
jupyter-server=2.17.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.5.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.28.0=pypi_0
jupyterlab-widgets=3.0.16=pypi_0
keyutils=1.6.3=hb9d3cd8_0
kiwisolver=1.4.9=pypi_0
krb5=1.21.3=h659f571_0
kubernetes=34.1.0=pypi_0
lark=1.3.1=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.45=default_hbd61a6d_105
libabseil=20240722.0=cxx17_hbbce691_4
libboost=1.86.0=hed09d94_4
libcurl=8.18.0=h4e3cde8_0
libedit=3.1.20250104=pl5321h7949ede_0
libev=4.33=hd590300_2
libexpat=2.7.3=hecca717_0
libffi=3.5.2=h9ec8514_0
libgcc=15.2.0=he0feb66_16
libgcc-ng=15.2.0=h69a702a_16
libgomp=15.2.0=he0feb66_16
liblzma=5.8.1=hb9d3cd8_2
libnghttp2=1.67.0=had1ee68_0
libnsl=2.0.1=hb9d3cd8_1
libsqlite=3.51.1=h0c1763c_1
libssh2=1.11.1=hcf80075_0
libstdcxx=15.2.0=h934c35e_16
libstdcxx-ng=15.2.0=hdf11a46_16
libuuid=2.41.3=h5347b49_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.3=pypi_0
matplotlib=3.10.8=pypi_0
matplotlib-inline=0.2.1=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.2.0=pypi_0
mongodb=6.0.16=h4be24d9_1
mpmath=1.3.0=pypi_0
nbclient=0.10.4=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.5=h2d0b736_3
nest-asyncio=1.6.0=pypi_0
networkx=3.6.1=pypi_0
notebook=7.5.1=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=2.4.0=pypi_0
oauthlib=3.3.1=pypi_0
openssl=3.6.0=h26f9b46_0
overrides=7.7.0=pypi_0
pandas=2.3.3=pypi_0
pandocfilters=1.5.1=pypi_0
panopticapi=0.1=pypi_0
parso=0.8.5=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.9.0=pypi_0
pillow=12.1.0=pypi_0
pip=25.3=pyh8b19718_0
platformdirs=4.5.1=pypi_0
prometheus-client=0.23.1=pypi_0
prompt-toolkit=3.0.52=pypi_0
psutil=7.2.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pycparser=2.23=pypi_0
pydantic=2.12.5=pypi_0
pydantic-core=2.41.5=pypi_0
pygments=2.19.2=pypi_0
pyparsing=3.3.1=pypi_0
python=3.11.14=hd63d673_2_cpython
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=4.0.0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.3=pypi_0
pyzmq=27.1.0=pypi_0
readline=8.3=h853b02a_0
referencing=0.37.0=pypi_0
regex=2025.11.3=pypi_0
requests=2.32.5=pypi_0
requests-oauthlib=2.0.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rfc3987-syntax=1.1.0=pypi_0
rich=14.2.0=pypi_0
rpds-py=0.30.0=pypi_0
rsa=4.9.1=pypi_0
rtree=1.4.1=pypi_0
safetensors=0.7.0=pypi_0
scikit-image=0.26.0=pypi_0
scikit-learn=1.8.0=pypi_0
scipy=1.16.3=pypi_0
send2trash=2.0.0=pypi_0
setuptools=80.9.0=pyhff2d567_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
snappy=1.2.2=h03e3b7b_1
sortedcontainers=2.4.0=pypi_0
soupsieve=2.8.1=pypi_0
stack-data=0.6.3=pypi_0
submitit=1.5.4=pypi_0
tabulate=0.9.0=pypi_0
termcolor=3.3.0=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2025.12.20=pypi_0
timm=1.0.24=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.13=noxft_ha0e22de_103
tokenizers=0.22.2=pypi_0
torchrun=0.2.9=pypi_0
tornado=6.5.4=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.57.3=pypi_0
typer=0.21.1=pypi_0
typing-inspection=0.4.2=pypi_0
tzdata=2025.3=pypi_0
tzlocal=5.3.1=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.40.0=pypi_0
wcwidth=0.2.14=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.9.0=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
widgetsnbextension=4.0.15=pypi_0
wrapt=2.0.1=pypi_0
yaml-cpp=0.8.0=h3f2d84a_0
yapf=0.43.0=pypi_0
zstd=1.5.7=hb78ec9c_6
